Move P-classes and repositories to own package
