+++ src/main/kotlin/no/nav/syfo/App.kt
@@ -9,7 +9,7 @@ import no.nav.syfo.api.apiModule
 import no.nav.syfo.application.IVurderingRepository
 import no.nav.syfo.application.service.ForhandsvarselService
 import no.nav.syfo.infrastructure.azuread.AzureAdClient
-import no.nav.syfo.infrastructure.database.VurderingRepository
+import no.nav.syfo.infrastructure.database.repository.VurderingRepository
 import no.nav.syfo.infrastructure.pdl.PdlClient
 import no.nav.syfo.infrastructure.database.applicationDatabase
 import no.nav.syfo.infrastructure.database.databaseModule
