+++ src/main/kotlin/no/nav/syfo/infrastructure/database/repository/PVarselPdf.kt
@@ -1,4 +1,4 @@
-package no.nav.syfo.infrastructure.database
+package no.nav.syfo.infrastructure.database.repository
 
 import java.time.OffsetDateTime
 import java.util.UUID
