+++ src/test/kotlin/no/nav/syfo/api/TestApiModule.kt
@@ -3,7 +3,7 @@ package no.nav.syfo.api
 import io.ktor.server.application.*
 import no.nav.syfo.ExternalMockEnvironment
 import no.nav.syfo.application.service.ForhandsvarselService
-import no.nav.syfo.infrastructure.database.VurderingRepository
+import no.nav.syfo.infrastructure.database.repository.VurderingRepository
 import no.nav.syfo.infrastructure.pdfgen.VarselPdfService
 import no.nav.syfo.infrastructure.veiledertilgang.VeilederTilgangskontrollClient
